refactor(vault): split notes.ts into focused files behind facade
Per .claude/plans/tech-debt-remediation.md Phase 3.1. The 1,546-line
notes.ts with 43 exports is split into 6 focused files. notes.ts stays as
a re-export facade so none of the 15 consumer files need updating.

- notes-io.ts (55 LOC): path utilities — getNotesDir, toAbsolutePath,
  toRelativePath, emitNoteEvent
- notes-crud.ts (733 LOC): CRUD + folders + utilities + import
  (createNote, updateNote, deleteNote, getNoteById, getFileById,
   getNoteByPath, noteExists, folder management, openExternal,
   revealInFinder, importFiles, all 14 exported types)
- notes-queries.ts (213 LOC): cross-note queries (listNotes,
  getTagsWithCounts, getNoteLinks, noteToListItem,
  extractAllLinkContexts) — pulled out to keep notes-crud under
  the 800-line pre-commit guardrail
- notes-rename.ts (189 LOC): renameNote, moveNote
- notes-versions.ts (244 LOC): snapshots + version history
  (createSnapshot, maybeCreateSignificantSnapshot, getVersionHistory,
   getVersion, restoreVersion + 3 threshold constants)
- notes.ts (13 LOC): facade re-exporting all 43 original exports

Type-only imports resolve the notes-crud ↔ notes-versions cycle. Existing
notes.test.ts (1,220 LOC, uses namespace import) runs unchanged.
